fix: Resolve remaining RimWorld 1.6 API compatibility errors
- Fix float to int cast in GameStateContext.cs (line 183)
- Replace PollutionGrid.GetTotalPercentWithinRadius with manual calculation
- Replace string.ToStringFull() with ToString()
- Replace GlowGrid.GameGlowAt with GroundGlowAt (RimWorld 1.6 API)
- Fix blueprint handling in PowerTools.cs (use CompForbiddable instead of Building_Blueprint)
- Stub Blueprint.WorkDone calculation (property removed in 1.6)
- Replace Room.PsychoActive with ProperRoom check

All 9 build errors resolved for RimWorld 1.6 compatibility.

+++ Source/RimMind/Chat/GameStateContext.cs
@@ -180,7 +180,7 @@ namespace RimMind.Chat
 
                     foreach (var letter in recentLetters)
                     {
-                        int minutesAgo = (currentTick - letter.arrivalTime) / (60 * 60);
+                        int minutesAgo = (int)((currentTick - letter.arrivalTime) / (60 * 60));
                         events.Add($"{letter.Label} ({minutesAgo}m ago)");
                     }
                 }
